# swagger.yml
openapi: 3.0.1
info:
  title: Photo Contest
  description: "This is Glints PTB1 Final Project.
    Github project at [https://github.com/ptbglints/photocontest](https://github.com/ptbglints/photocontest)
    For this sample, you can use the api key `special-key` to test the authorization filters."
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://thisurlwillbedynamicallychanged.io
...
paths:
...
  /users/{id}:
    get:
      tags:
        - User
      summary: Get specific user by id
      description: >
        A detailed description of the operation.
        Use markdown for rich text representation,
      operationId: getMany
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: "#/components/schemas/User"
...
  /users/profile/userid/{id}:
    get:
      tags:
        - Profile
      summary: Gets profile by userId
      # description:
      # operationId: getUserById
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
            example: 4
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /users/profile/:
    put:
      tags:
        - Profile
      summary: Update users own profile
      # description: This can only be done by the logged in user.
      # operationId: createUser
      requestBody:
        required: false
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                  example: 4
                name:
                  type: string
                  example: "Arman Maulalan"
                # address:
                #   type: string
                #   example: 'Jl. Pinang Kayu no.30 Bandung'
                # profilephoto:
                #   type: string
                # coverphoto:
                #   type: string
      responses:
        default:
          description: successful operation
          content: {}
      x-codegen-request-body-name: body
  /users/profile/username/{username}:
    get:
      tags:
        - Profile
      summary: Gets specific profile by Username
      # description:
      # operationId: getUserById
      parameters:
        - name: username
          in: path
          description: User name
          required: true
          schema:
            type: string
            example: "alice"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
components:
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userName:
          type: string
        email:
          type: string
        password:
          type: string
          minLength: 8
          maxLength: 254
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        name:
          type: string
        addres:
          type: string
        profilePhoto:
          type: string
        coverPhoto:
          type: string
        userId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...